Revert "Feature/my page"
