+++ src/components/CategoryNav.tsx
@@ -1,6 +1,5 @@
 import styled from "styled-components";
 import { useState } from "react";
-import { useNavigate } from "react-router-dom";
 
 const Wrap = styled.div`
   width: 100%;
@@ -19,13 +18,9 @@ const Li = styled.li`
   font-weight: 400;
   padding: 5px;
   border-radius: 10px;
-  background-color: #373e59;
-  cursor: pointer;
-  :hover {
-    color: #2196f3;
-  }
+  background-color: #187f7f;
 `;
-function CategoryNav({ view }: any) {
+function CategoryNav() {
   const [category, setCategory] = useState([
     "리그오브레전드",
     "롤토체스",
@@ -33,14 +28,11 @@ function CategoryNav({ view }: any) {
     "배틀그라운드",
     "로스트아크",
   ]);
-  const navigate = useNavigate();
   return (
     <Wrap>
       <Ul>
         {category.map((data, index) => (
-          <Li onClick={() => navigate(`/gamepage/${data}`)} key={index}>
-            {data}
-          </Li>
+          <Li key={index}>{data}</Li>
         ))}
       </Ul>
     </Wrap>
