+++ src/index.tsx
@@ -63,7 +63,9 @@ body {
   background-color: #373E59;
   color: whitesmoke;
   line-height: 1.2;
-  /* min-width: 1000px; */
+  overflow-x:hidden;
+  display: flex;
+  justify-content: center;
 }
 a {
   text-decoration:none;
