--- src/utils/gameCategory.tsx
@@ -1,22 +1,0 @@
-export const categorys = [
-  {
-    game: "리그 오브 레전드",
-    img: "https://kr-publish.s3.amazonaws.com/notice/0268f49d4d229bcee46589954838b252d9aa3f89.jpg",
-  },
-  {
-    game: "전략적 팀 전투 TFT",
-    img: "https://estnn.com/wp-content/uploads/2022/05/tftdragonlandsft.jpg",
-  },
-  {
-    game: "오버워치",
-    img: "https://image.aladin.co.kr/product/11734/84/letslook/8934978856_t10.jpg",
-  },
-  {
-    game: "배틀그라운드",
-    img: "https://thumb.mt.co.kr/06/2020/09/2020092817000468349_1.jpg/dims/optimize/",
-  },
-  {
-    game: "로스트아크",
-    img: "https://www.theguru.co.kr/data/photos/20220206/art_16443668651563_b158ed.jpg",
-  },
-];
